requirements.txt
faiss-cpu==1.7.4
langchain==0.0.197
openai==0.27.8
pypdf==3.9.1
python-dotenv==1.0.0
streamlit==1.23.1
streamlit-extras==0.2.7
tiktoken==0.4.0